Round vertex coordinates in a hash function to avoid floating point errors

+++ TombLib/TombLib/LevelData/Compilers/TombEngine/Structs.cs
@@ -169,7 +169,7 @@ namespace TombLib.LevelData.Compilers.TombEngine
 
         public override int GetHashCode()
         {
-            return HashCode.Combine(Position.X, Position.Y, Position.Z);
+            return HashCode.Combine((int)Position.X, (int)Position.Y, (int)Position.Z);
         }
     }
